Use size_t instead of int in isNameValid

+++ src/utils.c
@@ -71,8 +71,8 @@ int isNameValid(const char* name) {
     if(name == NULL || name[0] == '\0')
         return 0;
 
-    const int nameLength = strlen(name);
-    for(int i = 0; i < nameLength; i++)
+    const size_t nameLength = strlen(name);
+    for(size_t i = 0; i < nameLength; i++)
         if(!isalpha((unsigned char)name[i]))
             return 0;
